skillsb

+++ api/src/main/java/pl/kolendateam/dadcard/skills/dto/SkillsDTO.java
@@ -4,6 +4,6 @@ public class SkillsDTO {
 
     public String nameSkill;
     public boolean classSkill;
-    public double skillRank;
+    public int skillRank;
     
 }
